+++ cli/tests/test_status_command.rs
@@ -38,6 +38,11 @@ fn test_status_copies() {
     Working copy  (@) : rlvkpnrz c2fce842 (no description set)
     Parent commit (@-): qpvuntsm ebf799bc (no description set)
     [EOF]
+    ------- stderr -------
+    Auto-tracking 2 new files:
+    A copy-target
+    A rename-target
+    [EOF]
     ");
 }
 
@@ -88,6 +93,10 @@ fn test_status_ignored_gitignore() {
     Working copy  (@) : qpvuntsm af6b0e75 (no description set)
     Parent commit (@-): zzzzzzzz 00000000 (empty) (no description set)
     [EOF]
+    ------- stderr -------
+    Auto-tracking 1 new file:
+    A .gitignore
+    [EOF]
     ");
 
     let output = work_dir.run_jj(["status", "--ignored"]);
@@ -119,6 +128,11 @@ fn test_status_filtered() {
     A file_1
     Working copy  (@) : qpvuntsm 2f169edb (no description set)
     Parent commit (@-): zzzzzzzz 00000000 (empty) (no description set)
+    [EOF]
+    ------- stderr -------
+    Auto-tracking 2 new files:
+    A file_1
+    A file_2
     [EOF]
     ");