+++ CHANGELOG.md
@@ -29,6 +29,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * New option `--ignored` for `jj status` to have the list of ignored path
   currently in the working copy printed.
 
+* Print a message to indicate to the user when files starts to be tracked by
+  the auto-tracking system.
+
 ### Fixed bugs
 
 ## [0.39.0] - 2026-03-04